+++ src/Router.tsx
@@ -1,7 +1,17 @@
 import { createBrowserRouter } from 'react-router-dom';
-import { Home, Login, Search, Save, Example, SearchResult, KakaoCallBack } from '@/pages';
+import {
+  Home,
+  Login,
+  Search,
+  Save,
+  Example,
+  SearchResult,
+  KakaoCallBack,
+  MyPageContainer,
+  MyPage,
+  ProfileEdit,
+} from '@/pages';
 import App from '@/App';
-import MyPage from './pages/MyPage';
 
 const router = createBrowserRouter([
   {
@@ -26,7 +36,17 @@ const router = createBrowserRouter([
           },
           {
             path: 'my-page',
-            element: <MyPage />,
+            element: <MyPageContainer />,
+            children: [
+              {
+                path: '',
+                element: <MyPage />,
+              },
+              {
+                path: 'profile-edit',
+                element: <ProfileEdit />,
+              },
+            ],
           },
         ],
       },
