+++ src/pages/index.tsx
@@ -5,5 +5,19 @@ import Search from './Search';
 import Example from './Example';
 import SearchResult from './SearchResult';
 import KakaoCallBack from './KakaoCallBack';
+import MyPageContainer from './myPage/Container';
+import MyPage from './myPage/MyPage';
+import ProfileEdit from './myPage/ProfileEdit';
 
-export { Home, Login, Save, Search, Example, SearchResult, KakaoCallBack };
+export {
+  Home,
+  Login,
+  Save,
+  Search,
+  Example,
+  SearchResult,
+  KakaoCallBack,
+  MyPageContainer,
+  MyPage,
+  ProfileEdit,
+};
